smart-eco-tour-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai>=1.3.0
requests==2.31.0
numpy>=1.26.0
scikit-learn>=1.3.2
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
